Use logo template for logotheme logo
The logo for the logo theme is more representative of placing the logo
in a reasonable position. The titlepage is a special case. With this in
mind the logo template which is used with the \insertlogo command is the
appropriate size for the logobar theme, and the title has a special
template.