Fix #8 overfull vbox warning when compiling logobar theme
Beamer needs to have all elements defined on every slide, even if those
elements are transparent. The titlepage is an exception to this. Because
the headline was only defined for slides beyond the first, when the
extra element appeared it wasn't happy givinga overfull vbox warning.

This defines the logobar as transparent elements for the titleslide to
satiate the typesetter and get rid of the warnings. This also required
some adjustments to the slide title.