Create single examples directory
Put all the examples in a single directory to make it easier to build
all the examples with a single Makefile.

This also removes all the generated pdf files from the repository to
prevent the size ballooning as I continue to make changes. The idea is
to include the generated files as an embed in the README