Compile examples on travis when repo is updated (#10)
This configures travis-ci to build the repository when files are uploaded as a check that the compilation works. Additionally I am uploading the compiled files on tagged releases so updating the latest release just requires pushing a tag.

* Create single examples directory

Put all the examples in a single directory to make it easier to build
all the examples with a single Makefile.

This also removes all the generated pdf files from the repository to
prevent the size ballooning as I continue to make changes. The idea is
to include the generated files as an embed in the README

* Create travis configuration

This uses tectonic and biber 2.5 to compile the latex files to pdf on travis-ci. Tectonic was chosen because it is a very simple and quick install using conda, and it will automatically download the pacakges it requires from CTAN. Tectonic doesn't have support for biber, which is the recommended method of compiling a bibliography, so I have installed and run it separately.

* Include links to release versions in README

Now I am building the releases rather than including them in the
repository I have included links in the README to the respective files
which are representative of each theme element.